+++ src/app/search/SearchFilter.tsx
@@ -6,7 +6,6 @@ import { connect } from 'react-redux';
 import { RootState } from '../store/reducers';
 import { setSearchQuery } from '../shell/actions';
 import * as _ from 'lodash';
-import { toaster } from '../ngimport-more';
 import './search-filter.scss';
 import { destinyVersionSelector, currentAccountSelector } from '../accounts/reducer';
 import { SearchConfig, searchFilterSelector, searchConfigSelector } from './search-filters';
@@ -19,6 +18,7 @@ import { DimItem } from '../inventory/item-types';
 import { StoreServiceType } from '../inventory/store-types';
 import { loadingTracker } from '../shell/loading-tracker';
 import SearchFilterInput from './SearchFilterInput';
+import { showNotification } from '../notifications/notifications';
 
 const bulkItemTags = Array.from(itemTags) as any[];
 // t('Tags.TagItems')
@@ -99,18 +99,18 @@ class SearchFilter extends React.Component<Props, State> {
             // TODO: Gotta do this differently in react land
             item.locked = state;
           }
-          toaster.pop(
-            'success',
-            state
+          showNotification({
+            type: 'success',
+            title: state
               ? t('Filter.LockAllSuccess', { num: lockables.length })
               : t('Filter.UnlockAllSuccess', { num: lockables.length })
-          );
+          });
         } catch (e) {
-          toaster.pop(
-            'error',
-            state ? t('Filter.LockAllFailed') : t('Filter.UnlockAllFailed'),
-            e.message
-          );
+          showNotification({
+            type: 'error',
+            title: state ? t('Filter.LockAllFailed') : t('Filter.UnlockAllFailed'),
+            body: e.message
+          });
         } finally {
           // Touch the stores service to update state
           if (lockables.length) {