+++ docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Remember the last direction the infusion fuel finder was left in.
 * Remember the last option (equip or store) the "pull item" tool was left in.
+* Updated notification style. You can still click the notification to dismiss it.
+
+# 5.14.0 (2019-02-10)
 
 # 5.13.0 (2019-02-03)